+++ NEWS.md
@@ -5,6 +5,8 @@
 ### New features
 
 - Arithmetic operations between list columns are supported (#1589).
+- `polars_info()` shows the corresponding Python Polars version and the supported
+  Polars CompatLevel.
 
 ## polars 1.4.0